refactor message position
