+++ packages/types/src/message/common/base.ts
@@ -2,10 +2,10 @@
 import type { ILobeAgentRuntimeErrorType } from '@lobechat/model-runtime';
 import type { IPluginErrorType } from '@lobehub/chat-plugin-sdk';
 
-import { ErrorType } from '../fetch';
-import { GroundingSearch } from '../search';
+import { ErrorType } from '../../fetch';
+import { GroundingSearch } from '../../search';
+import { ChatToolPayload, MessageToolCall } from '../tools';
 import { ChatImageItem } from './image';
-import { ChatToolPayload, MessageToolCall } from './tools';
 
 /**
  * 聊天消息错误对象
