+++ packages/types/src/message/common/image.ts
