+++ packages/types/src/message/common/index.ts
@@ -1,0 +1,3 @@
+export * from './base';
+export * from './image';
+export * from './translate';
