+++ packages/types/src/message/index.ts
@@ -1,9 +1,6 @@
-export * from './base';
-export * from './chat';
-export * from './image';
-export * from './rag';
+export * from './common';
 export * from './tools';
-export * from './video';
+export * from './ui';
 
 export interface ModelRankItem {
   count: number;
